+++ alchemy.c
@@ -773,8 +773,8 @@ int mixer_use(int cn, int in) {
 
     endtime = ticker + duration;
 
-    *(signed long *)(it[in2].drdata) = endtime;
-    *(signed long *)(it[in2].drdata + 4) = ticker;
+    *(int *)(it[in2].drdata) = endtime;
+    *(int *)(it[in2].drdata + 4) = ticker;
 
     it[in2].carried = cn;
