+++ drvlib.c
@@ -1950,7 +1950,7 @@ int add_bonus_spell(int cn, int driver, int strength, int duration) {
 
     endtime = ticker + duration;
 
-    *(unsigned long *)(it[in].drdata) = endtime;
+    *(unsigned int *)(it[in].drdata) = endtime;
 
     it[in].carried = cn;
